requirements.txt
Pillow
Requests